UI Mostly Done

+++ ui/ui_utils.c
@@ -14,10 +14,10 @@ void gwemu_setUint8_to_register_widgets(GtkWidget **widgets, uint8_t value) {
 }
 
 //__attribute__((optimize("unroll-loops")))
-void gwemu_setUint16_to_register_widgets(GtkWidget **widgets, uint8_t value) {
+void gwemu_setUint16_to_register_widgets(GtkWidget **widgets, uint16_t value) {
     for (uint8_t i = 0; i < 16; i++) {
         gtk_image_set_from_icon_name(GTK_IMAGE(widgets[i]),
-                                     (value & (1U << i)) ? "gtk-yes" : "gtk-no",
+                                     (value & ( 1U << i)) ? "gtk-yes" : "gtk-no",
                                      GTK_ICON_SIZE_LARGE_TOOLBAR);
     }
 }
