+++ ui/ui_utils.h
@@ -8,6 +8,6 @@
 #define I8080SIM_UI_UTILS_H
 
 void gwemu_setUint8_to_register_widgets(GtkWidget **widgets, uint8_t value);
-void gwemu_setUint16_to_register_widgets(GtkWidget **widgets, uint8_t value);
+void gwemu_setUint16_to_register_widgets(GtkWidget **widgets, uint16_t value);
 
 #endif //I8080SIM_UI_UTILS_H
